Heretic, Hub, mix sound effects
Added Hub, mix sound effects capabilities and, also, posibility to chose to compile Doom or Heretic

+++ Libraries/STM32_USB_Host_Library/Core/Inc/usbh_ioreq.h
@@ -122,12 +122,12 @@ USBH_StatusTypeDef USBH_BulkSendData (USBH_HandleTypeDef *phost,
 
 USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost, 
                                        uint8_t             *buff, 
-                                       uint8_t             length,
+                                       uint16_t             length,
                                        uint8_t             hc_num);
 
 USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost, 
                                     uint8_t *buff, 
-                                    uint8_t length,
+                                    uint16_t length,
                                     uint8_t hc_num);
